Play-Command now does not require URL. Search is implemented into LoadSong Function which is used by Play-Command

+++ channel/channelmanager.py
@@ -13,6 +13,9 @@ class ChannelManager(commands.Cog):
 
     @commands.Cog.listener()
     async def on_voice_state_update(self, member, before, after):
+        if before.channel is None:
+            return
+
         if before.channel.name in self.chnls and len(before.channel.members) < 1:
             await before.channel.delete()
             del self.chnls[before.channel.name]
